tests imports refactor
