+++ tests/sparseml/onnx/optim/quantization/test_calibration.py
@@ -5,7 +5,7 @@ import onnx
 import pytest
 
 from sparseml.onnx.optim.quantization.calibration import CalibrationSession
-from tests.onnx.optim.quantization.helpers import (
+from tests.sparseml.onnx.optim.quantization.helpers import (
     make_tmp_onnx_file,
     onnx_conv_net,
     onnx_linear_net,
