+++ tests/sparseml/onnx/utils/test_graph_optimizer.py
@@ -6,7 +6,7 @@ import onnxruntime as rt
 import pytest
 
 from sparseml.onnx.utils.graph_optimizer import fold_conv_bns
-from tests.onnx.optim.quantization.helpers import make_tmp_onnx_file, onnx_conv_net
+from tests.sparseml.onnx.optim.quantization.helpers import make_tmp_onnx_file, onnx_conv_net
 
 
 def _model_has_conv_bn(model: onnx.ModelProto):
