+++ tests/sparseml/pytorch/models/classification/test_mobilenet.py
@@ -5,7 +5,7 @@ import pytest
 import torch
 
 from sparseml.pytorch.models import ModelRegistry, mobilenet
-from tests.pytorch.models.utils import compare_model
+from tests.sparseml.pytorch.models.utils import compare_model
 
 
 @pytest.mark.skipif(
