+++ tests/sparseml/pytorch/models/classification/test_resnet.py
@@ -22,7 +22,7 @@ from sparseml.pytorch.models import (
     resnext101,
     resnext152,
 )
-from tests.pytorch.models.utils import compare_model
+from tests.sparseml.pytorch.models.utils import compare_model
 
 
 @pytest.mark.skipif(
