+++ tests/sparseml/pytorch/models/external/test_torchvision.py
@@ -4,7 +4,7 @@ from typing import Callable, Union
 import pytest
 import torch
 from sparseml.pytorch.models import ModelRegistry
-from tests.pytorch.models.utils import compare_model
+from tests.sparseml.pytorch.models.utils import compare_model
 from torchvision import models as torchvision_models
