+++ tests/sparseml/pytorch/optim/test_analyzer_module.py
@@ -7,7 +7,7 @@ from torch.nn import Module
 from torchvision.models import resnet50
 
 from sparseml.pytorch.optim import ModuleAnalyzer
-from tests.pytorch.helpers import ConvNet, MLPNet
+from tests.sparseml.pytorch.helpers import ConvNet, MLPNet
 
 
 @pytest.mark.skipif(
