+++ tests/sparseml/pytorch/optim/test_modifier_lr.py
@@ -7,8 +7,8 @@ from torch.optim import SGD, Adam
 
 from sparseml.pytorch.optim import LearningRateModifier, SetLearningRateModifier
 from sparseml.pytorch.utils import get_optim_learning_rate
-from tests.pytorch.helpers import LinearNet
-from tests.pytorch.optim.test_modifier import (
+from tests.sparseml.pytorch.helpers import LinearNet
+from tests.sparseml.pytorch.optim.test_modifier import (
     ScheduledModifierTest,
     ScheduledUpdateModifierTest,
     test_epoch,
