+++ tests/sparseml/pytorch/optim/test_modifier_quantization.py
@@ -5,7 +5,7 @@ import torch
 from torch.nn import Conv2d, Linear
 
 from sparseml.pytorch.optim import QuantizationModifier
-from tests.pytorch.helpers import (
+from tests.sparseml.pytorch.helpers import (
     ConvNet,
     LinearNet,
     create_optim_sgd,
@@ -13,7 +13,7 @@ from tests.pytorch.helpers import (
     test_loss,
     test_steps_per_epoch,
 )
-from tests.pytorch.optim.test_modifier import ScheduledModifierTest
+from tests.sparseml.pytorch.optim.test_modifier import ScheduledModifierTest
 
 
 try:
