+++ tests/sparseml/pytorch/utils/test_benchmarker.py
@@ -6,7 +6,7 @@ import pytest
 import torch
 
 from sparseml.pytorch.utils import BatchBenchmarkResults, ModuleBenchmarker
-from tests.pytorch.helpers import MLPNet
+from tests.sparseml.pytorch.helpers import MLPNet
 
 
 @pytest.mark.skipif(
