+++ tests/sparseml/tensorflow_v1/optim/test_mask_pruning.py
@@ -22,7 +22,7 @@ from sparseml.tensorflow_v1.utils import (
     get_op_input_var,
     tf_compat,
 )
-from tests.tensorflow_v1.helpers import conv_net, mlp_net
+from tests.sparseml.tensorflow_v1.helpers import conv_net, mlp_net
 
 
 @pytest.mark.skipif(
