+++ tests/sparseml/tensorflow_v1/utils/test_variable.py
@@ -10,7 +10,7 @@ from sparseml.tensorflow_v1.utils import (
     get_prunable_ops,
     tf_compat,
 )
-from tests.tensorflow_v1.helpers import conv_net, mlp_net
+from tests.sparseml.tensorflow_v1.helpers import conv_net, mlp_net
 
 
 @pytest.mark.skipif(
